move EncodedCbor struct to separate mod
