+++ rust/vote-tx-v2/src/gen_tx/mod.rs
@@ -11,7 +11,7 @@ mod vote;
 pub use builder::GeneralizedTxBuilder;
 use coset::CborSerializable;
 pub use event_map::{EventKey, EventMap};
-use minicbor::{data::Tag, Decode, Decoder, Encode, Encoder};
+use minicbor::{Decode, Decoder, Encode, Encoder};
 pub use tx_body::{TxBody, VoterData};
 pub use vote::{Choice, Proof, PropId, Vote};
 
@@ -31,14 +31,6 @@ where
     /// `signature` field
     signature: coset::CoseSign,
 }
-
-/// An encoded CBOR struct, CBOR tag 24.
-#[derive(Debug, Clone, PartialEq)]
-pub struct EncodedCbor<T>(T)
-where T: for<'a> Cbor<'a>;
-
-/// encoded-cbor CBOR tag <https://www.iana.org/assignments/cbor-tags/cbor-tags.xhtml/>.
-const ENCODED_CBOR_TAG: u64 = 24;
 
 /// `GeneralizedTx` array struct length
 const GENERALIZED_TX_LEN: u64 = 2;
@@ -109,39 +101,6 @@ where
     }
 }
 
-impl<T> Decode<'_, ()> for EncodedCbor<T>
-where T: for<'a> Cbor<'a>
-{
-    fn decode(d: &mut Decoder<'_>, (): &mut ()) -> Result<Self, minicbor::decode::Error> {
-        let tag = d.tag()?;
-        if ENCODED_CBOR_TAG != tag.as_u64() {
-            return Err(minicbor::decode::Error::message(format!(
-                "tag value must be: {ENCODED_CBOR_TAG}, provided: {}",
-                tag.as_u64(),
-            )));
-        }
-        let cbor_bytes = d.bytes()?.to_vec();
-        let cbor = T::from_bytes(&cbor_bytes).map_err(minicbor::decode::Error::message)?;
-        Ok(Self(cbor))
-    }
-}
-
-impl<T> Encode<()> for EncodedCbor<T>
-where T: for<'a> Cbor<'a>
-{
-    fn encode<W: minicbor::encode::Write>(
-        &self, e: &mut minicbor::Encoder<W>, (): &mut (),
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        e.tag(Tag::new(ENCODED_CBOR_TAG))?;
-        let cbor_bytes = self
-            .0
-            .to_bytes()
-            .map_err(minicbor::encode::Error::message)?;
-        e.bytes(&cbor_bytes)?;
-        Ok(())
-    }
-}
-
 /// Reads CBOR bytes from the decoder and returns them as bytes.
 fn read_cbor_bytes(d: &mut Decoder<'_>) -> Result<Vec<u8>, minicbor::decode::Error> {
     let start = d.position();
@@ -169,7 +128,7 @@ mod tests {
     use test_strategy::proptest;
 
     use super::*;
-    use crate::uuid::Uuid;
+    use crate::{encoded_cbor::EncodedCbor, uuid::Uuid};
 
     type ChoiceT = Vec<u8>;
     type ProofT = Vec<u8>;
