+++ rust/vote-tx-v2/src/lib.rs
@@ -6,6 +6,7 @@
 use anyhow::anyhow;
 use minicbor::{Decode, Decoder, Encode, Encoder};
 
+pub mod encoded_cbor;
 pub mod gen_tx;
 pub mod public_tx;
 pub mod uuid;
